requirements.txt
appnope==0.1.0
asn1crypto==0.22.0
bcrypt==3.1.3
beautifulsoup4==4.6.0
bleach==2.0.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
cryptography==1.9
decorator==4.0.11
Django==1.10.7
django-appconf==1.0.2
django-contrib-comments==1.8.0
django-modeltranslation==0.12.1
django-post-office==2.0.8
django-widget-tweaks==1.4.1
ecdsa==0.13
Fabric3==1.13.1.post1
filebrowser-safe==0.4.7
future==0.16.0
geopy==1.11.0
grappelli-safe==0.4.6
html5lib==0.999999999
httplib2==0.10.3
icalendar==3.11.4
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
jsonfield==2.0.2
Mezzanine==4.2.3
mezzanine-agenda==0.1.2
-e git+https://github.com/georgeyk/mezzanine-instagram-gallery.git@a4543063634410488b0ca2b068cf44c7b5d151ea#egg=mezzanine_instagram_gallery
oauthlib==2.0.2
olefile==0.44
paramiko==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
ptyprocess==0.5.2
pyasn1==0.2.3
pycparser==2.17
pycrypto==2.6.1
Pygments==2.2.0
PyNaCl==1.1.2
python-dateutil==2.6.0
python-instagram==1.3.2
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
simplegeneric==0.8.1
simplejson==3.11.1
six==1.10.0
traitlets==4.3.2
typing==3.6.1
tzlocal==1.4
urllib3==1.21.1
wcwidth==0.1.7
webencodings==0.5.1
